+++ Magic/src/main/java/com/elmakers/mine/bukkit/entity/EntityData.java
@@ -679,7 +679,6 @@ public class EntityData
             }
         }
         if (spawned == null && type != null && type != EntityType.PLAYER) {
-            boolean previous = controller.isDisableSpawnReplacement();
             controller.setDisableSpawnReplacement(true);
             try {
                 SpawnedEntityExtraData spawnedEntity = null;
@@ -695,7 +694,7 @@ public class EntityData
             } catch (Exception ex) {
                 org.bukkit.Bukkit.getLogger().log(Level.WARNING, "Error restoring entity type " + getType() + " at " + getLocation(), ex);
             }
-            controller.setDisableSpawnReplacement(previous);
+            controller.setDisableSpawnReplacement(false);
         }
         if (spawned != null) {
             try {