+++ Magic/src/main/java/com/elmakers/mine/bukkit/magic/MagicController.java
@@ -7751,7 +7751,7 @@ public class MagicController implements MageController, ChunkLoadListener {
     }
 
     @Override
-    public void setDisableSpawnReplacement(boolean disable) {
+    public synchronized void setDisableSpawnReplacement(boolean disable) {
         if (disable) {
             disableSpawnReplacement++;
         } else {